Refactor service plugins
